Handle Rate Limiting
The Unsplash API has a rate limit of 50 requests per hour. In order to have this demo be public, we need to ensure that we load backup data in the event the rate limit is reached, rather than just showing a blank page. We also shuffle the default data so that refreshing the page changes the layout every time.